Use explicit conversions for xcoord, ycoord and ship size

+++ src/main/scala/com/lambdarat/navalcombat/engine/BoardEngine.scala
@@ -1,27 +1,27 @@
 package com.lambdarat.navalcombat.engine
 
 import com.lambdarat.navalcombat.core.*
+import com.lambdarat.navalcombat.core.XCoord.*
+import com.lambdarat.navalcombat.core.YCoord.*
 import com.lambdarat.navalcombat.core.Rotation.*
 import com.lambdarat.navalcombat.core.Cell.*
 import com.lambdarat.navalcombat.core.given
 
-import scala.language.implicitConversions
-
 object BoardEngine:
 
-  private def validCoords(x: XCoord, y: YCoord): Boolean = x >= 0 && x < 10 && y >= 0 && y < 10
+  private def validCoords(x: XCoord, y: YCoord): Boolean = x.toInt >= 0 && x.toInt < 10 && y.toInt >= 0 && y.toInt < 10
 
   extension (board: Board)
 
     def get(x: XCoord, y: YCoord): Option[Cell] =
-      Option.when(validCoords(x, y))(board.cells(x)(y))
+      Option.when(validCoords(x, y))(board.cells(x.toInt)(y.toInt))
 
     def update(x: XCoord, y: YCoord, value: Cell): Board =
       Option
         .when(validCoords(x, y)) {
-          val row          = board.cells(x)
-          val updatedRow   = row.updated(y, value)
-          val updatedCells = board.cells.updated(x, updatedRow)
+          val row          = board.cells(x.toInt)
+          val updatedRow   = row.updated(y.toInt, value)
+          val updatedCells = board.cells.updated(x.toInt, updatedRow)
 
           board.copy(cells = updatedCells)
         }
@@ -33,7 +33,7 @@ object BoardEngine:
       val maybeCanPlace = Option.when(validCoords(x, y)) {
         rotation match
           case Vertical =>
-            (0 until ship.size)
+            (0 until ship.size.toInt)
               .forall(shipY =>
                 validCoords(x, y + shipY) && board.get(x, y + shipY).fold(false) {
                   case Unknown => true
@@ -41,7 +41,7 @@ object BoardEngine:
                 }
               )
           case Horizontal =>
-            (0 until ship.size)
+            (0 until ship.size.toInt)
               .forall(shipX =>
                 validCoords(x + shipX, y) && board.get(x + shipX, y).fold(false) {
                   case Unknown => true
@@ -57,11 +57,11 @@ object BoardEngine:
       if canPlace(ship, rotation, x, y) then
         rotation match
           case Vertical =>
-            (0 until ship.size).foldLeft(board) { case (oldBoard, shipY) =>
+            (0 until ship.size.toInt).foldLeft(board) { case (oldBoard, shipY) =>
               oldBoard.update(x, y + shipY, Floating(ship))
             }
           case Horizontal =>
-            (0 until ship.size).foldLeft(board) { case (oldBoard, shipX) =>
+            (0 until ship.size.toInt).foldLeft(board) { case (oldBoard, shipX) =>
               oldBoard.update(x + shipX, y, Floating(ship))
             }
       else board
